backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
asyncpg
aiosqlite
psycopg[binary]
alembic
pydantic-settings
PyJWT
pandas
openpyxl
scikit-learn
papermill
nbformat
python-multipart
cryptography
kagglehub
httpx
chromadb
tiktoken
pytest
pytest-asyncio